requirements.txt
gitastroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-bootstrap4==0.0.7
django-braces==1.13.0
django-countries==5.3.2
django-crispy-forms==1.7.2
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
PyJWT==1.7.0
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
